requirements.txt
# Weaviate集成依赖
weaviate-client>=4.0.0
langchain-weaviate>=0.0.1

# 基础依赖（与主requirements.txt相同）
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20
openai>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
aiohttp>=3.9.0
requests>=2.31.0